♻️(worker) Introduce an `IPool` interface
